feat(canvas): deepen proof and split render hotspots (#3799)
Agent Canvas v2 now proves recovery redirects, DOM snapshots, console diagnostics, pixel and screenshot markers, and controlled click/type/probe controls through the repeatable ops-chat canvas proof. The first render hotspot split moves chat markdown, dashboard agent canvas, dashboard deploy lifecycle, and ops deploy form handling into focused modules while keeping route contracts stable.

Constraint: Preserve existing /ops/chat and /ops/deploy marker contracts for PR #3761

Constraint: Keep this as extraction plus proof hardening, not a UI redesign

Rejected: Leave proof as artifact-only | it would not catch broken recovery redirects or runtime contract markers

Rejected: Continue growing lib.rs and ops_dashboard_shell.rs | those files are already the main maintainability hotspots

Confidence: high

Scope-risk: moderate

Reversibility: clean

Directive: Continue moving dashboard/gateway render surfaces into focused modules before adding new route complexity

Tested: git diff --check

Tested: scripts/dev/test-ops-chat-canvas-proof.sh

Tested: scripts/dev/test-whats-missing.sh

Tested: CARGO_TARGET_DIR=/tmp/rust_pi-codex-issue-3758-deploy-stop-process-lifecycle-fast-validate-full-target RUST_MIN_STACK=16777216 cargo test -p tau-dashboard-ui functional_spec_3799 -- --nocapture

Tested: CARGO_TARGET_DIR=/tmp/rust_pi-codex-issue-3758-deploy-stop-process-lifecycle-fast-validate-full-target RUST_MIN_STACK=16777216 cargo test -p tau-gateway functional_spec_3799 -- --nocapture

Tested: /opt/homebrew/bin/timeout 1800s env RUST_MIN_STACK=16777216 scripts/dev/fast-validate.sh --full

Not-tested: cargo mutants --in-diff
