+++ crates/tau-gateway/src/gateway_openresponses.rs
@@ -72,6 +72,7 @@ mod openresponses_entry_handler;
 mod openresponses_execution_handler;
 mod ops_chat_canvas;
 mod ops_dashboard_shell;
+mod ops_deploy_panel;
 mod ops_harness_memory_graph;
 mod ops_harness_proposals;
 mod ops_shell_controls;
@@ -184,14 +185,14 @@ use openresponses_execution_handler::execute_openresponses_request;
 use ops_dashboard_shell::{
     handle_ops_dashboard_channel_action, handle_ops_dashboard_chat_new,
     handle_ops_dashboard_chat_send, handle_ops_dashboard_chat_send_get,
-    handle_ops_dashboard_control_action, handle_ops_dashboard_deploy_stop,
-    handle_ops_dashboard_deploy_submit, handle_ops_dashboard_harness_artifact,
+    handle_ops_dashboard_control_action, handle_ops_dashboard_harness_artifact,
     handle_ops_dashboard_harness_artifact_view, handle_ops_dashboard_harness_create_mission_draft,
     handle_ops_dashboard_harness_proposal_action, handle_ops_dashboard_harness_proposal_diff,
     handle_ops_dashboard_harness_run_benchmark, handle_ops_dashboard_harness_start_mission,
     handle_ops_dashboard_memory_create, handle_ops_dashboard_session_detail_reset,
     handle_ops_dashboard_sessions_branch, render_tau_ops_dashboard_shell_for_route,
 };
+use ops_deploy_panel::{handle_ops_dashboard_deploy_stop, handle_ops_dashboard_deploy_submit};
 use ops_harness_memory_graph::{
     append_ops_harness_memory_graph_lineage, collect_ops_harness_memory_graph_lineage,
 };
